requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
prisma==0.11.0
psycopg2-binary==2.9.9
redis==5.0.1
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
boto3==1.34.23
django-storages==1.14.2
stripe==7.10.0
prometheus-client==0.19.0
sentry-sdk==1.39.2
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
python-decouple==3.8
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.2.0
drf-spectacular==0.27.0
black==23.12.1
flake8==7.0.0
isort==5.13.2